+++ src/modules/location-huds/locations/afterword-acres/index.js
@@ -1,0 +1,13 @@
+import { addHudStyles } from '@utils';
+import folkloreForest from '../../shared/folklore-forest';
+
+import regionStyles from '../../shared/folklore-forest/styles.css';
+
+/**
+ * Initialize the module.
+ */
+export default async () => {
+  addHudStyles([regionStyles]);
+
+  folkloreForest();
+};