+++ src/modules/location-huds/locations/epilogue-falls/index.js
@@ -1,0 +1,12 @@
+import folkloreForest from '../../shared/folklore-forest';
+
+import regionStyles from '../../shared/folklore-forest/styles.css';
+
+/**
+ * Initialize the module.
+ */
+export default async () => {
+  addHudStyles([regionStyles]);
+
+  folkloreForest();
+};